+++ ImperatorToCK3/Imperator/Provinces/Province.cs
@@ -24,7 +24,7 @@ internal sealed partial class Province : IIdentifiable<ulong> {
 	public ulong? HolySiteId { get; set; } = null;
 	public ulong? HoldingOwnerId { get; set; } = null;
 	public uint BuildingCount { get; set; } = 0;
-	public double CivilizationValue { get; set; } = 0;
+	public float CivilizationValue { get; set; } = 0;
 
 	public Province(ulong id) {
 		Id = id;
